编码

+++ sjt-services/sjt-business-web/src/main/java/com/sjt/wechat/vo/req/pay/WxPayUnifiedRequestVO.java
@@ -114,7 +114,7 @@ public class WxPayUnifiedRequestVO {
     @Element(
         name = "total_fee"
     )
-    private Long totalFee;
+    private Integer totalFee;
 
     /**
      * 终端IP
